feat: add optional optimized but incompatible power function
This introduces an optimized power function. The benefit of this is that
we can control the blower independently from the flame.
There are two scenarios where this could be useful:

1. A user wants to run the blower after the flame has already been turned off.
   The official app always turns off both blower and flame when the power
   off button is used.

2. A user wants to turn on the flame at a flame height other than 6.
   The official app always turns the flame height to 6 when turning on.

This method of control is incompatible with the official app, as it will
start showing invalid state if connected to a fireplace that has been
powered on or off using this method. Changing any setting from within
the app will return the app to its normal state though.

By default the compatible control mechanism is used, so this change is
backwards compatible.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>